requirements.txt
Flask==3.1.3
flask-cors==6.0.2
python-dotenv==1.2.1
gunicorn==23.0.0
